requirements.txt
ccxt==4.3.89
pandas==2.2.2
pandas-ta==0.3.14b
numpy==1.26.4
python-dotenv==1.0.1
aiohttp==3.9.5
loguru==0.7.2
colorama==0.4.6
